requirements.txt
pyvis
langchain-openai
python-dotenv
streamlit
pydantic
langchain-core
PyPDF2
